fix: 2nd window bug
