+++ app/lib/constants/app/app_links.dart
@@ -7,5 +7,6 @@ class AppLinks {
       'https://docs.google.com/document/d/1DWY-ZNv7cUA6egxDZKYu0e8qz7z-yHT2KncyYAo5NHU/edit?pli=1';
   static const String reportBug =
       'https://github.com/Basiiii/Profit-Taker-Analytics/issues/new/choose';
+  static const String leaderboard = 'https://stats.profit-taker.com/';
   static const basiPaypal = "https://www.paypal.com/paypalme/basigraphics";
 }
