+++ app/lib/screens/settings/settings_screen.dart
@@ -7,13 +7,13 @@ import 'package:profit_taker_analyzer/services/input/action_keys.dart';
 import 'package:profit_taker_analyzer/widgets/dialogs/show_onboarding_dialog.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:profit_taker_analyzer/widgets/dialogs/donation_dialog.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 
 import 'ui/general_section.dart';
 import 'ui/links_section.dart';
 import 'ui/about_section.dart';
 import 'ui/key_config_section.dart';
+import 'ui/account_section.dart';
+import 'ui/donate_section.dart';
 
 /// A StatefulWidget that displays the settings screen of the application.
 ///
@@ -150,7 +150,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
   void _onSignIn(BuildContext context) {
     showDialog(
       context: context,
-      builder: (context) => _SignInDialog(),
+      builder: (context) => SignInDialog(),
     );
   }
 
@@ -196,201 +196,4 @@ class _SettingsScreenState extends State<SettingsScreen> {
       ),
     );
   }
-}
-
-class _SignInDialog extends StatefulWidget {
-  @override
-  State<_SignInDialog> createState() => _SignInDialogState();
-}
-
-class _SignInDialogState extends State<_SignInDialog> {
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  bool _isLoading = false;
-  String? _error;
-
-  Future<void> _signInWithEmail() async {
-    setState(() {
-      _isLoading = true;
-      _error = null;
-    });
-    try {
-      final response = await Supabase.instance.client.auth.signInWithPassword(
-        email: _emailController.text.trim(),
-        password: _passwordController.text,
-      );
-      if (!mounted) return;
-      if (response.user == null) {
-        setState(() {
-          _error = 'Invalid credentials';
-        });
-      } else {
-        Navigator.of(context).pop();
-      }
-    } catch (e) {
-      if (!mounted) return;
-      setState(() {
-        _error = e.toString();
-      });
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isLoading = false;
-        });
-      }
-    }
-  }
-
-  Future<void> _signInWithDiscord() async {
-    setState(() {
-      _isLoading = true;
-      _error = null;
-    });
-    try {
-      await Supabase.instance.client.auth.signInWithOAuth(
-        OAuthProvider.discord,
-        redirectTo: 'pta://auth-callback',
-      );
-      if (!mounted) return;
-      Navigator.of(context).pop();
-    } catch (e) {
-      if (!mounted) return;
-      setState(() {
-        _error = e.toString();
-      });
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isLoading = false;
-        });
-      }
-    }
-  }
-
-  void _forgotPassword() async {
-    final url =
-        Uri.parse('https://stats.profit-taker.com/auth/forgot-password');
-    if (await canLaunchUrl(url)) {
-      await launchUrl(url, mode: LaunchMode.externalApplication);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: Text('Sign In'),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          TextField(
-            controller: _emailController,
-            decoration: InputDecoration(labelText: 'Email'),
-            keyboardType: TextInputType.emailAddress,
-          ),
-          TextField(
-            controller: _passwordController,
-            decoration: InputDecoration(labelText: 'Password'),
-            obscureText: true,
-          ),
-          Align(
-            alignment: Alignment.centerRight,
-            child: TextButton(
-              onPressed: _forgotPassword,
-              child: Text('Forgot password?'),
-            ),
-          ),
-          if (_error != null) ...[
-            SizedBox(height: 8),
-            Text(_error!, style: TextStyle(color: Colors.red)),
-          ],
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: _isLoading ? null : () => Navigator.of(context).pop(),
-          child: Text('Cancel'),
-        ),
-        ElevatedButton(
-          onPressed: _isLoading ? null : _signInWithEmail,
-          child: _isLoading
-              ? SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(strokeWidth: 2))
-              : Text('Sign In'),
-        ),
-        OutlinedButton.icon(
-          onPressed: _isLoading ? null : _signInWithDiscord,
-          icon: Icon(Icons.discord),
-          label: Text('Discord'),
-        ),
-      ],
-    );
-  }
-}
-
-SettingsSection buildAccountSection(
-  BuildContext context,
-  User? user,
-  String? username,
-  VoidCallback onSignIn,
-  VoidCallback onSignUp,
-  VoidCallback onLogout,
-) {
-  if (user != null && username != null) {
-    return SettingsSection(
-      title: Text(FlutterI18n.translate(context, "settings.account.title")),
-      tiles: [
-        SettingsTile(
-          title: Text(FlutterI18n.translate(
-              context, "settings.account.logged_in_as",
-              translationParams: {"username": username})),
-          leading: const Icon(Icons.account_circle),
-        ),
-        SettingsTile(
-          title:
-              Text(FlutterI18n.translate(context, "settings.account.logout")),
-          leading: const Icon(Icons.logout),
-          onPressed: (_) => onLogout(),
-        ),
-      ],
-    );
-  } else {
-    return SettingsSection(
-      title: Text(FlutterI18n.translate(context, "settings.account.title")),
-      tiles: [
-        SettingsTile(
-          title:
-              Text(FlutterI18n.translate(context, "settings.general.sign_in")),
-          leading: const Icon(Icons.login),
-          onPressed: (_) => onSignIn(),
-        ),
-        SettingsTile(
-          title:
-              Text(FlutterI18n.translate(context, "settings.general.sign_up")),
-          leading: const Icon(Icons.person_add),
-          onPressed: (_) => onSignUp(),
-        ),
-      ],
-    );
-  }
-}
-
-SettingsSection buildDonateSection(BuildContext context) {
-  return SettingsSection(
-    title: Text(FlutterI18n.translate(context, "donate.title")),
-    tiles: [
-      SettingsTile(
-        title: Text(FlutterI18n.translate(context, "donate.button")),
-        leading: const Icon(Icons.favorite),
-        onPressed: (_) => showDonationDialog(
-          context,
-          FlutterI18n.translate(context, "donate.title"),
-          FlutterI18n.translate(context, "donate.main"),
-          FlutterI18n.translate(context, "donate.button"),
-          FlutterI18n.translate(context, "common.ok"),
-        ),
-      ),
-    ],
-  );
 }
